Wall Run Started
Lower gravity when onWall is true. ISSUE: Jump when onGround and onWall goes way to high, need to have it so upward velocity gets lowerd or zeroed out when onWall is true.

Also, Need to make it so player is somehow connected to wall and movement away or jumping disconnects player from wall. Have so movement is only possible on Wall's orientation?