Camera Roates Bassed on Wall Completed
Camera will rotate on the z asix to the left if collision with a wall is to the player's right. The opposite will happen if the wall is on the player's left.

This will also change while the player is on the wall and rotating the camera/player. If player is on the wall and wall is to the right, the camera z axis will rotate to the left. While the player is still on the wall and the player rotates character so that the wall is now detected to be on the player's left, the camera will rotate to the right.

Also, if player is on the wall and sys cannot detect if wall is on the right or left, camera will move to normal position (z-axis rotation is set to 0).

Made it so there is smooth transition between each rotation of the z-axis (right, left, normal).