Wall Run Implemented
Able to move along Wall by utilizing the cross product of the collided with wall normal and Vector3.up. (wallNormal X Vector3.up). It gives the direction vector that is along the surface of the wall.

Had to create a new WallMove method to handle wall movement.

Need to test curved surfaces.