Fixed Camera Tilt Issue
Issue: Camera tilt spins when transitioning between states due to -angle to positive angle change and vice versa.

Fix: Make rotation of the camera a animation. Animations for Idle, Tilt Right/Left, and Tilt Right/Left Stay created and applied to the Anim Controller.

There was an Issue doing it this way. The animation controller cancelled out the hard code x-axis rotation for looking up and down. The fix for which was to create a new empty game object and give it the Camera Control script for looking up and down. Make this object the child of the player model, and have the Main Camera be a child to that empty game object.