Goal, Music, and Speed Hud implemented
Goal implemented and has collision and trigger which works. Doesn't actiavte anything yet. It just knows when the player goes through the trigger.

Music implemented and plays a playlist of music when in a level. Specific tracks can be called on.

Game now displays the xz speed of the player as an integer.