Created a Start object and a Starting Script, Updated Goal and Timer Script
Created a Starting point/gate that starts the timer on the hud. If the player passes through the gate, the timer will reset.

Updated Goal Script to stop the timer when the player passes thru.

Made it so the models for Goal and Start have 3D Text that let the player know which is which.

Created a Shader and Text Material that HW Andy uses to make it so the 3D Text isn't overlayed on top of all other 3D objects in the scene. (See Research Links)