Continued Work on Character Movement
Able to crouch while in the air, and the only thing affected is the player scale.

Added a upward/downward force when colliding with a wall to adjust player y velocity. Like adding friction to a wall and make it seem more like the player is attaching to the wall and compensating.

Made it so checking if player is in the air to update status is more prioritized when a player is sliding. Status would say sliding instead of being in the air before.